+++ src/Musurp/Neo/Cypher/AbstractComponent.php
@@ -12,26 +12,24 @@ declare(strict_types=1);
 namespace Musurp\Neo\Cypher;
 
 /**
- * An abstract query component.
- *
- * This class represents a base for anything in the Cypher syntax tree.
+ * An abstract component representing syntax elements.
  */
 abstract class AbstractComponent
 {
     /**
-     * Return the expression as string.
+     * Compile the expression to string.
      *
      * @return string
      */
-    abstract public function toString(): string;
+    abstract public function compile(): string;
 
     /**
-     * Allow magic return of the expression as string.
+     * A magic method to allow all components to be cast to string.
      *
      * @return string
      */
     public function __toString(): string
     {
-        return $this->toString();
+        return $this->compile();
     }
 }