+++ src/Musurp/Neo/Cypher/Component/Expression/Operator/AbstractLogicalOperator.php
@@ -44,7 +44,7 @@ abstract class AbstractLogicalOperator extends AbstractExpressionComponent
     /**
      * {@inheritdoc}
      */
-    public function toString(): string
+    public function compile(): string
     {
         $glue = sprintf(' %s ', $this->getOperator());