+++ src/Musurp/Neo/Cypher/Tests/Unit/Component/Expression/Operator/Logical/AndLogicalOperatorTest.php
@@ -40,7 +40,7 @@ class AndLogicalOperatorTest extends TestCase
 (TRUE AND FALSE)
 CYPHER;
 
-        self::assertEquals($cypher, $operator->toString());
+        self::assertEquals($cypher, $operator->compile());
     }
 
     /**
@@ -65,7 +65,7 @@ CYPHER;
 (TRUE AND (TRUE AND 'hello'))
 CYPHER;
 
-        self::assertEquals($cypher, $operator->toString());
+        self::assertEquals($cypher, $operator->compile());
     }
 
     /**
@@ -99,7 +99,7 @@ CYPHER;
 (((TRUE AND 'foo') AND 'bar') AND (FALSE AND (4 AND 7.4)))
 CYPHER;
 
-        self::assertEquals($cypher, $operator->toString());
+        self::assertEquals($cypher, $operator->compile());
     }
 
     /**
@@ -138,6 +138,6 @@ CYPHER;
 (((TRUE AND 'foo') AND 'bar') AND (FALSE AND (4 AND 7.4) AND FALSE) AND (2 AND 'hammer'))
 CYPHER;
 
-        self::assertEquals($cypher, $operator->toString());
+        self::assertEquals($cypher, $operator->compile());
     }
 }